+++ CHANGES.md
@@ -7,6 +7,7 @@
 ## Upcoming Changes
 
  - Detect unapplied Redact annotations as bad redactions ([issue](https://github.com/freelawproject/x-ray/issues/194))
+ - Skip white-on-white redactions to reduce false positives ([issue](https://github.com/freelawproject/x-ray/issues/196))
 
 ## Current Version